pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api"
version = "0.0.1"
description = "FastAPI backend of LostMa data"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "fastapi [standard] >= 0.113.0 , < 0.114.0",
    "pydantic >= 2.7.0 , < 3.0.0",
    "duckdb >= 1.2.1",
    "heurist-api>=0.2.3",
    "pydantic-settings>=2.9.1",
    "kuzu>=0.10.0",
]

[project.optional-dependencies]
dev = ["black>=24.1.0", "isort>=5.13.0", "pytest"]

[tool.hatch.build.targets.wheel]
packages = ["app", "database"]

[tool.pytest.ini_options]
# Ignore the DB tests when running full pytest
addopts = ["--import-mode=importlib", "-p no:warnings", "--ignore=tests/db"]
testspaths = ["tests/app"]
